LLM-Powered API Tester & Documentation Generator/docutest-ai/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.14.3
pydantic==2.6.4
python-multipart==0.0.9
